Update README.md
"BankingReports is a PHP-based application that simplifies the process of generating and managing banking reports for businesses. It allows workers to submit banking reports, including bank slips, and provides administrators with a centralized platform to track and analyze the data. With an intuitive dashboard and powerful reporting capabilities, BankingReports streamlines financial management and enhances transparency. Easily manage your banking reports and gain valuable insights with BankingReports."